colin's comments
