+++ .gitignore
@@ -1,5 +1,5 @@
 /target
-
+Cargo.lock
 *.png
 *.html
 *.DS_Store
