+++ src/examples.rs
@@ -456,34 +456,3 @@ pub fn brownian_bridge_plotter(display: Display, start_price: f64, end_price: f6
     //plot
     transparent_plot(Some(vec![curve1, curve2]), None, axes, title, display);
 }
-#[allow(unused)]
-/// Plot imported csv data for single column csv's
-pub fn csv_plotter(display: Display, file_path: &str) {
-    let title = "\\text{csv Data}".to_string();
-    // Import the data from the csv file
-    let value = import_price_from_csv(file_path).unwrap();
-    // Use a parameterization of the curves to build them
-    let t_start = 0.0;
-    let number_of_points = value.len();
-    let t_end = number_of_points as f64;
-    let t = linspace(t_start, t_end, number_of_points).collect::<Vec<f64>>();
-    // build curve
-    let curve = Curve {
-        x_coordinates: t.clone(),
-        y_coordinates: value,
-        design: ElementDesign {
-            color: Color::Green,
-            color_slot: MAIN_COLOR_SLOT,
-            emphasis: Emphasis::Light,
-        },
-        name: Some(String::from("\\text{csv data}")),
-    };
-    // build plot axes
-    let axes = Axes {
-        x_label: String::from("event number"),
-        y_label: String::from("value"),
-        bounds: (vec![0.0, t_end], vec![0.0, 3000.0]),
-    };
-    //plot
-    transparent_plot(Some(vec![curve]), None, axes, title, display);
-}
