+++ src/main.rs
@@ -2,6 +2,9 @@ mod design;
 mod examples;
 mod functions;
 mod plot;
+mod file_handler;
+#[allow(unused)]
+use file_handler::*;
 use design::DisplayMode;
 use examples::*;
 use plot::*;
